part I of the big refactor!
