+++ package.json
@@ -1,31 +1,31 @@
 {
   "name": "ipfsd-ctl",
   "version": "0.27.3",
-  "description": "simple controls for an ipfs node",
+  "description": "Spawn IPFS Daemons, JS or Go",
   "main": "src/index.js",
   "scripts": {
     "lint": "aegir lint",
-    "coverage": "COVERAGE=true aegir coverage --timeout 50000",
+    "docs": "aegir docs",
     "test": "aegir test -t node -t browser --no-cors",
     "test:node": "aegir test -t node",
     "test:browser": "aegir test -t browser --no-cors",
-    "docs": "aegir docs",
-    "release": "aegir release -t node",
-    "release-minor": "aegir release --type minor -t node",
-    "release-major": "aegir release --type major -t node",
+    "release": "aegir release",
+    "release-minor": "aegir release --type minor",
+    "release-major": "aegir release --type major",
+    "coverage": "COVERAGE=true aegir coverage --timeout 50000",
     "coverage-publish": "aegir coverage -u"
   },
   "browser": {
-    "./src/utils/create-repo-nodejs.js": "./src/utils/create-repo-browser.js",
-    "./src/daemon-node.js": false,
-    "./src/remote-node/routes.js": false,
-    "./src/exec.js": false,
     "hapi": false,
     "glob": false,
     "fs": false,
     "joi": false,
     "stream": "readable-stream",
-    "http": "stream-http"
+    "http": "stream-http",
+    "./src/repo/create-nodejs.js": "./src/utils/create-browser.js",
+    "./src/daemon-factory.js": false,
+    "./src/remote/server.js": false,
+    "./src/exec.js": false
   },
   "engines": {
     "node": ">=6.0.0",
