--- src/remote-node/index.js
@@ -1,9 +1,0 @@
-'use strict'
-
-const Server = require('./server')
-const RemoteController = require('./client')
-
-module.exports = {
-  Server,
-  RemoteController
-}
