+++ src/remote/daemon-client.js
@@ -20,7 +20,7 @@ function createApi (apiAddr, gwAddr) {
   return api
 }
 
-class Node {
+class DaemonClient {
   constructor (baseUrl, id, apiAddr, gwAddrs) {
     this.baseUrl = baseUrl
     this._id = id
@@ -260,60 +260,4 @@ class Node {
   }
 }
 
-class RemoteFactory {
-  constructor (opts) {
-    opts = opts || {}
-    if (!opts.host) {
-      opts.host = 'localhost'
-    }
-
-    if (!opts.port) {
-      opts.port = 9999
-    }
-
-    if (typeof opts.host === 'number') {
-      opts.port = opts.host
-      opts.host = 'localhost'
-    }
-
-    this.port = opts.port
-    this.host = opts.host
-    this.type = opts.type || 'go'
-
-    if (this.type === 'proc') {
-      throw new Error(`'proc' is not allowed in remote mode`)
-    }
-
-    this.baseUrl = `${opts.secure ? 'https://' : 'http://'}${this.host}:${this.port}`
-  }
-
-  spawn (opts, cb) {
-    if (typeof opts === 'function') {
-      cb = opts
-      opts = {}
-    }
-
-    opts = opts || {}
-    request
-      .post(`${this.baseUrl}/spawn`)
-      .send({ opts, type: this.type })
-      .end((err, res) => {
-        if (err) {
-          return cb(new Error(err.response ? err.response.body.message : err))
-        }
-
-        const apiAddr = res.body.api ? res.body.api.apiAddr : ''
-        const gatewayAddr = res.body.api ? res.body.api.gatewayAddr : ''
-
-        const node = new Node(
-          this.baseUrl,
-          res.body.id,
-          apiAddr,
-          gatewayAddr)
-
-        cb(null, node)
-      })
-  }
-}
-
-module.exports = RemoteFactory
+module.exports = DaemonClient
